get JniBridgeSpec working
