+++ src/test/scala/chiseltest/simulator/jni/JniBridgeSpec.scala
@@ -1,17 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package chiseltest.tests
+package chiseltest.simulator.jni
 
-import org.scalatest._
-
-import chisel3._
-import chisel3.experimental.BundleLiterals._
 import chiseltest._
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import chiseltest.simulator.jni.JniAPI
-
-import java.io.File
 
 
 // object BridgeLib {
@@ -19,11 +12,11 @@ import java.io.File
 //   System.load(soPath.toString())
 // }
 
-class BridgeLib extends AnyFlatSpec with ChiselScalatestTester with Matchers {
+class JniBridgeSpec extends AnyFlatSpec with ChiselScalatestTester with Matchers {
   behavior of "Testers2"
 
   it should "be able to load so through bridge library and call functions" in {
-    println("BridgeLib start")
+    //println("BridgeLib start")
     val soPath = os.pwd / "src" / "main" / "resources" / "jni" / "VFoo"
 
     val vfoo_id = JniAPI.load_so(soPath.toString())
@@ -54,7 +47,7 @@ class BridgeLib extends AnyFlatSpec with ChiselScalatestTester with Matchers {
     assert(output_1 == 30)
     assert(output2_1 == 40)
 
-    JniAPI.call_finish(vfoo_id, ptr);
-    JniAPI.call_finish(vfoo_id2, ptr);
+    JniAPI.call_finish(vfoo_id, ptr)
+    JniAPI.call_finish(vfoo_id2, ptr)
   }
 }
